requirements.txt
python-dotenv
requests
openai==0.2
paddleocr>=2.0.1